Rename folder redux to features
