+++ src/features/generalSlice.ts
