+++ src/pages/PrintersEdit.tsx
@@ -1,4 +1,4 @@
-import { usePrinters } from '../redux/hooks'
+import { usePrinters } from '../hooks'
 
 import { useEffect, useRef, useState } from 'react'
 import { Link, useNavigate, useParams } from 'react-router-dom'
